Jogosultság ellenőrző kiépítése minden oldalra
- Active-Directory jogosultság ellenőrzése, oldal és tartalom típustól függően
- Karbantartásra és onnan visszalépés lehetőségének kialakítása
- Nyomtatási lábléc módosítása
