+++ ISZR.csproj
@@ -8,8 +8,8 @@
 		<GenerateDocumentationFile>False</GenerateDocumentationFile>
 		<SignAssembly>False</SignAssembly>
 		<Version></Version>
-		<AssemblyVersion>23.2.9.7</AssemblyVersion>
-		<FileVersion>23.2.9.7</FileVersion>
+		<AssemblyVersion>23.2.10.12</AssemblyVersion>
+		<FileVersion>23.2.10.12</FileVersion>
 	</PropertyGroup>
 
 	<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
