+++ Views/Administrator/Index.cshtml
@@ -2,7 +2,7 @@
 
 @{
     Layout = "~/Views/Shared/_AdminLayout.cshtml";
-    ViewData["Title"] = "ISZR Karbantartása";
+    ViewData["Title"] = "Igénylések karbantartása";
 }
 
 <div class="container-fluid">
