+++ Views/Home/Dashboard.cshtml
@@ -1,4 +1,6 @@
-﻿@{
+﻿@using ISZR.Components;
+
+@{
 	ViewData["Title"] = "Irányítópult";
 }
 
@@ -30,45 +32,48 @@
 				</div>
 			</div>
 		</div>
-		<div class="col-12 col-lg-8">
-			<div class="card">
-				<div class="card-header bg-info">
-					<h5 class="card-title">Igényléseid eloszlása</h5>
-				</div>
+		@if (Account.IsUgyintezo())
+		{
+			<div class="col-12 col-lg-8">
+				<div class="card">
+					<div class="card-header bg-info">
+						<h5 class="card-title">Igényléseid eloszlása</h5>
+					</div>
 
-				<div class="card-body">
-					<div class="row">
-						<div class="col-12">
-							@{
-								int done = (int)Math.Round((double)(100 * ViewBag.Done) / ViewBag.All);
-								int inProgress = (int)Math.Round((double)(100 * ViewBag.InProgress) / ViewBag.All);
-								int denied = (int)Math.Round((double)(100 * ViewBag.Denied) / ViewBag.All);
-							}
-							<div class="progress-group">
-								Jováhagyott igényléseid
-								<span class="float-right"><b>@ViewBag.Done darab</b> / @ViewBag.All darab</span>
-								<div class="progress progress-sm">
-									<div class="progress-bar bg-success" style="width: @done%"></div>
+					<div class="card-body">
+						<div class="row">
+							<div class="col-12">
+								@{
+									int done = (int)Math.Round((double)(100 * ViewBag.Done) / ViewBag.All);
+									int inProgress = (int)Math.Round((double)(100 * ViewBag.InProgress) / ViewBag.All);
+									int denied = (int)Math.Round((double)(100 * ViewBag.Denied) / ViewBag.All);
+								}
+								<div class="progress-group">
+									Jováhagyott igényléseid
+									<span class="float-right"><b>@ViewBag.Done darab</b> / @ViewBag.All darab</span>
+									<div class="progress progress-sm">
+										<div class="progress-bar bg-success" style="width: @done%"></div>
+									</div>
 								</div>
-							</div>
-							<div class="progress-group">
-								Folyamatban lévő igényléseid
-								<span class="float-right"><b>@ViewBag.InProgress darab</b> / @ViewBag.All darab</span>
-								<div class="progress progress-sm">
-									<div class="progress-bar bg-warning" style="width: @inProgress%"></div>
+								<div class="progress-group">
+									Folyamatban lévő igényléseid
+									<span class="float-right"><b>@ViewBag.InProgress darab</b> / @ViewBag.All darab</span>
+									<div class="progress progress-sm">
+										<div class="progress-bar bg-warning" style="width: @inProgress%"></div>
+									</div>
 								</div>
-							</div>
-							<div class="progress-group">
-								Elutasított igényléseid
-								<span class="float-right"><b>@ViewBag.Denied darab</b> / @ViewBag.All darab</span>
-								<div class="progress progress-sm">
-									<div class="progress-bar bg-danger" style="width: @denied%"></div>
+								<div class="progress-group">
+									Elutasított igényléseid
+									<span class="float-right"><b>@ViewBag.Denied darab</b> / @ViewBag.All darab</span>
+									<div class="progress progress-sm">
+										<div class="progress-bar bg-danger" style="width: @denied%"></div>
+									</div>
 								</div>
 							</div>
 						</div>
 					</div>
 				</div>
 			</div>
-		</div>
+		}
 	</div>
 </div>
