+++ Views/Shared/_AdminNavbar.cshtml
@@ -1,18 +1,25 @@
 ﻿@using System.Text.RegularExpressions;
 
 <nav class="main-header navbar navbar-expand navbar-white navbar-light">
-    <!-- Leftside -->
-    <ul class="navbar-nav">
-        <li class="nav-item">
-            <a class="nav-link" data-widget="pushmenu" href="#" role="button"><i class="fas fa-bars"></i></a>
-        </li>
-    </ul>
+	<!-- Leftside -->
+	<ul class="navbar-nav">
+		<li class="nav-item">
+			<a class="nav-link" data-widget="pushmenu" href="#" role="button"><i class="fas fa-bars"></i></a>
+		</li>
+	</ul>
 
-    <!-- Right side -->
-    <ul class="navbar-nav ml-auto">
-        <li class="nav-item">
-            <a class="nav-link" data-widget="fullscreen" href="#" role="button">
-                <i class="fas fa-expand-arrows-alt"></i>
-            </a>
-        </li>
+	<!-- Right side -->
+	<ul class="navbar-nav ml-auto">
+		<li class="nav-item">
+			<a class="nav-link" data-widget="fullscreen" href="#" role="button">
+				<i class="fas fa-expand-arrows-alt"></i>
+			</a>
+		</li>
+		<li class="nav-item">
+			<a class="nav-link btn btn-danger text-white" asp-area="" asp-controller="Home" asp-action="Dashboard">
+				<i class="fas fa-home"></i>
+				<span>Vissza az igénylő rendszerbe</span>
+			</a>
+		</li>
+	</ul>
 </nav>
