+++ Views/Shared/_Footer.cshtml
@@ -1,6 +1,10 @@
-﻿<footer class="main-footer font-weight-normal d-print-none">
-	&copy; 2021-@DateTime.Now.Year. Az alkalmazást készítette <a class="text-info font-weight-bold" href="https://github.com/Xentinus"> <i class="fab fa-github"></i> Béla Kellner</a>.
-
+﻿<footer class="main-footer font-weight-normal">
+	<div class="float-left">
+		&copy; 2021-@DateTime.Now.Year.
+		<span class="d-print-none">
+			Az alkalmazást készítette <a class="text-info font-weight-bold" href="https://github.com/Xentinus"> <i class="fab fa-github"></i> Béla Kellner</a>.
+		</span>
+	</div>
 	<div class="float-right d-none d-sm-inline">
 		Informatikai Szolgálati jegy Rendszer <span class="text-info font-weight-bold">
 			@System.Reflection.Assembly.GetExecutingAssembly().GetName().Version
