+++ Views/Shared/_Navbar.cshtml
@@ -1,25 +1,35 @@
-﻿@using System.Text.RegularExpressions;
+﻿@using ISZR.Components;
+@using System.Text.RegularExpressions;
 
 <nav class="main-header navbar navbar-expand navbar-white navbar-light">
-    <!-- Leftside -->
-    <ul class="navbar-nav">
-        <li class="nav-item">
-            <a class="nav-link" data-widget="pushmenu" href="#" role="button"><i class="fas fa-bars"></i></a>
-        </li>
-    </ul>
+	<!-- Leftside -->
+	<ul class="navbar-nav">
+		<li class="nav-item">
+			<a class="nav-link" data-widget="pushmenu" href="#" role="button"><i class="fas fa-bars"></i></a>
+		</li>
+	</ul>
 
-    <!-- Right side -->
-    <ul class="navbar-nav ml-auto">
-        <li class="nav-item">
-            <a class="nav-link" data-widget="fullscreen" href="#" role="button">
-                <i class="fas fa-expand-arrows-alt"></i>
-            </a>
-        </li>
-        <li class="nav-item">
-            <a class="nav-link" asp-area="" asp-controller="Home" asp-action="Settings">
-                <i class="fas fa-user-cog"></i>
-                <span>Felhasználói profil</span>
-            </a>
-        </li>
-    </ul>
+	<!-- Right side -->
+	<ul class="navbar-nav ml-auto">
+		<li class="nav-item">
+			<a class="nav-link" data-widget="fullscreen" href="#" role="button">
+				<i class="fas fa-expand-arrows-alt"></i>
+			</a>
+		</li>
+		<li class="nav-item">
+			<a class="nav-link btn btn-info text-white" asp-area="" asp-controller="Home" asp-action="Settings">
+				<i class="fas fa-user-cog"></i>
+				<span>Felhasználói profil</span>
+			</a>
+		</li>
+		@if (Account.IsAdmin())
+		{
+			<li class="nav-item ml-2">
+				<a class="nav-link btn btn-danger text-white" asp-area="" asp-controller="Administrator" asp-action="Index">
+					<i class="fas fa-toolbox"></i>
+					<span>Karbantartás</span>
+				</a>
+			</li>
+		}
+	</ul>
 </nav>
